Refactor metadata defaults and relocate tests
